+++ Lead-Client-Source/EffectLib/Type.cpp
@@ -13,8 +13,8 @@ BOOL GetTokenTimeEventFloat(CTextFileLoader & rTextFileLoader, const char * c_sz
 	DWORD dwIndex = 0;
 	for (DWORD i = 0; i < pTokenVector->size(); i+=2, ++dwIndex)
 	{
-		pTimeEventTableFloat->at(dwIndex).m_fTime = atof(pTokenVector->at(i).c_str());
-		pTimeEventTableFloat->at(dwIndex).m_Value = atof(pTokenVector->at(i+1).c_str());
+		pTimeEventTableFloat->at(dwIndex).m_fTime = static_cast<float>(atof(pTokenVector->at(i).c_str()));
+		pTimeEventTableFloat->at(dwIndex).m_Value = static_cast<float>(atof(pTokenVector->at(i+1).c_str()));
 	}
 
 	return TRUE;
